# Conflicts:
#	dockerfile
#	pom.xml
#	src/main/java/com/sr/common/TelephoneCheck.java
#	src/main/java/com/sr/controller/DownloadController.java
#	src/main/java/com/sr/controller/UploadController.java
#	src/main/java/com/sr/controller/UserController.java
#	src/main/java/com/sr/enunn/StatusEnum.java
#	src/main/java/com/sr/service/UserService.java
#	src/main/java/com/sr/service/impl/UserServiceImpl.java
#	src/main/resources/application-prod.yaml
#	src/main/resources/application-test.yaml
#	src/test/java/com/sr/EntityMapConvertTest.java

+++ src/main/java/com/sr/dao/UserMapper.java
@@ -16,7 +16,7 @@ public interface UserMapper {
 
     int insert(User record);
 
-    int insertSelective(User record);
+    long insertSelective(User record);
 
     List<User> selectByExample(UserExample example);
